+++ hahomematic/platforms/generic/number.py
@@ -5,16 +5,12 @@ See https://www.home-assistant.io/integrations/number/.
 """
 from __future__ import annotations
 
-import logging
-
-from hahomematic.const import HM_VALUE, HmPlatform
-from hahomematic.platforms.entity import CallParameterCollector, ParameterT
+from hahomematic.const import HmPlatform
+from hahomematic.platforms.entity import InputParameterT, ParameterT
 from hahomematic.platforms.generic.entity import GenericEntity
 
-_LOGGER = logging.getLogger(__name__)
 
-
-class BaseNumber(GenericEntity[ParameterT]):
+class BaseNumber(GenericEntity[ParameterT, InputParameterT]):
     """
     Implementation of a number.
 
@@ -24,60 +20,48 @@ class BaseNumber(GenericEntity[ParameterT]):
     _attr_platform = HmPlatform.NUMBER
 
 
-class HmFloat(BaseNumber[float]):
+class HmFloat(BaseNumber[float, float | str]):
     """
     Implementation of a Float.
 
     This is a default platform that gets automatically generated.
     """
 
-    async def send_value(
-        self,
-        value: float,
-        collector: CallParameterCollector | None = None,
-        do_validate: bool = True,
-    ) -> None:
-        """Set the value of the entity."""
-        if (
-            value is not None and self._attr_min <= float(value) <= self._attr_max
-        ) or not do_validate:
-            await super().send_value(value=value, collector=collector)
-        elif self._attr_special:
-            if [sv for sv in self._attr_special.values() if value == sv[HM_VALUE]]:
-                await super().send_value(value=value, collector=collector)
-        else:
-            _LOGGER.warning(
-                "NUMBER.FLOAT failed: Invalid value: %s (min: %s, max: %s, special: %s)",
-                value,
-                self._attr_min,
-                self._attr_max,
-                self._attr_special,
-            )
+    def _prepare_value_for_sending(self, value: float | str, do_validate: bool = True) -> float:
+        """Prepare value before sending."""
+        if not do_validate or (
+            value is not None
+            and isinstance(value, float)
+            and self._attr_min <= float(value) <= self._attr_max
+        ):
+            return float(value)
+        if self._attr_special and isinstance(value, str) and value in self._attr_special:
+            return float(self._attr_special[value])
+        raise ValueError(
+            f"NUMBER.FLOAT failed: Invalid value: {value} (min: {self._attr_min}, "
+            f"max: {self._attr_max}, special:{self._attr_special})"
+        )
 
 
-class HmInteger(BaseNumber[int]):
+class HmInteger(BaseNumber[int, int | str]):
     """
     Implementation of an Integer.
 
     This is a default platform that gets automatically generated.
     """
 
-    async def send_value(
-        self, value: int, collector: CallParameterCollector | None = None, do_validate: bool = True
-    ) -> None:
-        """Set the value of the entity."""
-        if (
-            value is not None and self._attr_min <= int(value) <= self._attr_max
-        ) or not do_validate:
-            await super().send_value(value=value, collector=collector)
-        elif self._attr_special:
-            if [sv for sv in self._attr_special.values() if value == sv[HM_VALUE]]:
-                await super().send_value(value=value, collector=collector)
-        else:
-            _LOGGER.warning(
-                "NUMBER.INT failed: Invalid value: %s (min: %s, max: %s, special: %s)",
-                value,
-                self._attr_min,
-                self._attr_max,
-                self._attr_special,
-            )
+    def _prepare_value_for_sending(self, value: int | str, do_validate: bool = True) -> int:
+        """Prepare value before sending."""
+        if not do_validate or (
+            value is not None
+            and isinstance(value, int)
+            and self._attr_min <= int(value) <= self._attr_max
+        ):
+            return int(value)
+        if self._attr_special and isinstance(value, str) and value in self._attr_special:
+            return int(self._attr_special[value])
+
+        raise ValueError(
+            f"NUMBER.INT failed: Invalid value: {value} (min: {self._attr_min}, "
+            f"max: {self._attr_max}, special:{self._attr_special})"
+        )