+++ hahomematic/platforms/generic/select.py
@@ -5,17 +5,12 @@ See https://www.home-assistant.io/integrations/select/.
 """
 from __future__ import annotations
 
-import logging
-
 from hahomematic.const import HmPlatform
-from hahomematic.platforms.entity import CallParameterCollector
 from hahomematic.platforms.generic.entity import GenericEntity
 from hahomematic.platforms.support import value_property
 
-_LOGGER = logging.getLogger(__name__)
 
-
-class HmSelect(GenericEntity[int | str]):
+class HmSelect(GenericEntity[int | str, int | str]):
     """
     Implementation of a select entity.
 
@@ -31,22 +26,15 @@ class HmSelect(GenericEntity[int | str]):
             return self._attr_value_list[int(self._attr_value)]
         return str(self._attr_default)
 
-    async def send_value(
-        self, value: int | str, collector: CallParameterCollector | None = None
-    ) -> None:
-        """Set the value of the entity."""
+    def _prepare_value_for_sending(self, value: int | str, do_validate: bool = True) -> int | str:
+        """Prepare value before sending."""
         # We allow setting the value via index as well, just in case.
-        if isinstance(value, int) and self._attr_value_list:
-            if 0 <= value < len(self._attr_value_list):
-                await super().send_value(value=value, collector=collector)
-        elif self._attr_value_list:
-            if value in self._attr_value_list:
-                await super().send_value(
-                    value=self._attr_value_list.index(value), collector=collector
-                )
-        else:
-            _LOGGER.warning(
-                "Value not in value_list for %s/%s",
-                self.name,
-                self.unique_identifier,
-            )
+        if (
+            isinstance(value, int)
+            and self._attr_value_list
+            and 0 <= value < len(self._attr_value_list)
+        ):
+            return value
+        if self._attr_value_list and value in self._attr_value_list:
+            return self._attr_value_list.index(value)
+        raise ValueError(f"Value not in value_list for {self.name}/{self.unique_identifier}")